Refactor base package name of databind project
