+++ example/lwjgl3-opengl/src/main/java/com/labymedia/ultralight/lwjgl3/opengl/support/ViewContextProvider.java
@@ -20,8 +20,8 @@
 package com.labymedia.ultralight.lwjgl3.opengl.support;
 
 import com.labymedia.ultralight.UltralightView;
-import com.labymedia.ultralight.context.ContextProviderFactory;
-import com.labymedia.ultralight.context.ContextProvider;
+import com.labymedia.ultralight.databind.context.ContextProvider;
+import com.labymedia.ultralight.databind.context.ContextProviderFactory;
 import com.labymedia.ultralight.javascript.JavascriptContextLock;
 import com.labymedia.ultralight.javascript.JavascriptValue;
