+++ ultralight-java-databind/src/main/java/com/labymedia/ultralight/databind/DatabindConfiguration.java
@@ -17,13 +17,13 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package com.labymedia.ultralight;
+package com.labymedia.ultralight.databind;
 
-import com.labymedia.ultralight.context.ContextProviderFactory;
-import com.labymedia.ultralight.cache.JavascriptClassCache;
-import com.labymedia.ultralight.cache.NaiveJavascriptClassCache;
-import com.labymedia.ultralight.call.HeuristicMethodChooser;
-import com.labymedia.ultralight.call.MethodChooser;
+import com.labymedia.ultralight.databind.cache.JavascriptClassCache;
+import com.labymedia.ultralight.databind.cache.NaiveJavascriptClassCache;
+import com.labymedia.ultralight.databind.call.HeuristicMethodChooser;
+import com.labymedia.ultralight.databind.call.MethodChooser;
+import com.labymedia.ultralight.databind.context.ContextProviderFactory;
 
 /**
  * Databind configuration.
