+++ ultralight-java-databind/src/main/java/com/labymedia/ultralight/databind/DatabindJavascriptExplicitAPI.java
@@ -17,11 +17,11 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package com.labymedia.ultralight;
+package com.labymedia.ultralight.databind;
 
+import com.labymedia.ultralight.databind.utils.JavascriptConversionUtils;
 import com.labymedia.ultralight.javascript.*;
 import com.labymedia.ultralight.javascript.*;
-import com.labymedia.ultralight.utils.JavascriptConversionUtils;
 
 import java.util.ArrayList;
 import java.util.List;
