+++ ultralight-java-databind/src/main/java/com/labymedia/ultralight/databind/api/InjectJavascriptContext.java
@@ -17,9 +17,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package com.labymedia.ultralight.api;
+package com.labymedia.ultralight.databind.api;
 
-import com.labymedia.ultralight.Databind;
+import com.labymedia.ultralight.databind.Databind;
 import com.labymedia.ultralight.javascript.JavascriptContext;
 
 import java.lang.annotation.ElementType;
