+++ ultralight-java-databind/src/main/java/com/labymedia/ultralight/databind/cache/JavascriptClassCache.java
@@ -17,7 +17,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package com.labymedia.ultralight.cache;
+package com.labymedia.ultralight.databind.cache;
 
 import com.labymedia.ultralight.javascript.JavascriptClass;
