+++ ultralight-java-databind/src/main/java/com/labymedia/ultralight/databind/call/HeuristicMethodChooser.java
@@ -17,12 +17,12 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package com.labymedia.ultralight.call;
+package com.labymedia.ultralight.databind.call;
 
-import com.labymedia.ultralight.api.InjectJavascriptContext;
+import com.labymedia.ultralight.databind.api.InjectJavascriptContext;
+import com.labymedia.ultralight.databind.utils.JavascriptConversionUtils;
 import com.labymedia.ultralight.javascript.JavascriptObject;
 import com.labymedia.ultralight.javascript.JavascriptValue;
-import com.labymedia.ultralight.utils.JavascriptConversionUtils;
 
 import java.lang.reflect.Executable;
 import java.lang.reflect.Modifier;
