+++ ultralight-java-databind/src/main/java/com/labymedia/ultralight/databind/context/ContextProviderFactory.java
@@ -17,9 +17,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package com.labymedia.ultralight.context;
+package com.labymedia.ultralight.databind.context;
 
-import com.labymedia.ultralight.DatabindConfiguration;
+import com.labymedia.ultralight.databind.DatabindConfiguration;
 import com.labymedia.ultralight.javascript.JavascriptValue;
 
 /**
